\name{WhitelawCowan-package}
\alias{WhitelawCowan-package}
\alias{WhitelawCowan}
\docType{package}
\title{
Functions and parameters needed to simulate 2D Whitelaw and Cowan
}
\description{
An implementation with a radius of activity and scaled total activity.
}
\details{
\tabular{ll}{
Package: \tab WhitelawCowan\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-10\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Catherine Cutts
}
\references{
Whitelaw and Cowan (1981).
}

\keyword{ package }

\examples{
source(system.file('runme.R', package='WhitelawCowan'))
}
